break apart pipeline-manager
